[refactor] replace main dashboard page for counters page
